BREAKING CHANGE: replace dropdown to new dropdown
