--- src/components/NewDropdown/index.ts
@@ -1,4 +1,0 @@
-export { Dropdown } from './Dropdown'
-export { DropdownTrigger } from './DropdownTrigger'
-export { DropdownContent } from './DropdownContent'
-export { DropdownControllableContent } from './DropdownControllableContent'
